Fix non-functional MTP.
This requires a minor patch to the kernel:

  https://github.com/ianmacd/z3s/commit/3ab4fec69869822273a029067e4c22863c722f37

and a kernel config change to avoid a reproducible spontaneous reboot
after 220 seconds of uptime in TWRP:

  https://github.com/ianmacd/z3s/commit/7f6a8e199147f4770c2902fd20e3926474c9007b

The default TWRP init.rc is now used in favour of Samsung's recovery
init.rc, and a custom USB configuration is applied.

Additionally, MTP no longer needs to be disabled to allow adb to work.